refactor: move extension methods to separate folder
